+++ tomcat/src/main/java/org/apache/coyote/controller/Controller.java
@@ -1,8 +1,8 @@
-package org.apache.coyote;
+package org.apache.coyote.controller;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
-import org.apache.coyote.http11.Request;
+import org.apache.coyote.http11.request.Request;
 import org.apache.coyote.http11.Response;
 
 public interface Controller {
