+++ tomcat/src/main/java/org/apache/coyote/controller/RegisterController.java
@@ -1,12 +1,14 @@
-package org.apache.coyote;
+package org.apache.coyote.controller;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
+import org.apache.coyote.http11.request.HttpMethod;
+import org.apache.coyote.http11.HttpStatus;
 import org.apache.coyote.exception.DuplicateAccountRegisterException;
-import org.apache.coyote.http11.Request;
-import org.apache.coyote.http11.RequestBody;
+import org.apache.coyote.http11.request.Request;
+import org.apache.coyote.http11.request.RequestBody;
 import org.apache.coyote.http11.Response;
 
 public class RegisterController implements Controller {
