+++ tomcat/src/main/java/org/apache/coyote/controller/StaticFileController.java
@@ -1,8 +1,9 @@
-package org.apache.coyote;
+package org.apache.coyote.controller;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
-import org.apache.coyote.http11.Request;
+import org.apache.coyote.http11.HttpStatus;
+import org.apache.coyote.http11.request.Request;
 import org.apache.coyote.http11.Response;
 
 public class StaticFileController implements Controller{
