+++ tomcat/src/main/java/org/apache/coyote/http11/HttpStatus.java
@@ -1,4 +1,4 @@
-package org.apache.coyote;
+package org.apache.coyote.http11;
 
 public enum HttpStatus {
     OK(200, "OK"),
