+++ tomcat/src/main/java/org/apache/coyote/http11/request/QueryParams.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
 
 import java.util.HashMap;
 import java.util.Map;
