+++ tomcat/src/test/java/nextstep/org/apache/coyote/ControllerContainerTest.java
@@ -2,11 +2,11 @@ package nextstep.org.apache.coyote;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import org.apache.coyote.Controller;
-import org.apache.coyote.ControllerContainer;
-import org.apache.coyote.HttpMethod;
-import org.apache.coyote.StaticFileController;
-import org.apache.coyote.http11.Request;
+import org.apache.coyote.controller.Controller;
+import org.apache.coyote.controller.ControllerContainer;
+import org.apache.coyote.http11.request.HttpMethod;
+import org.apache.coyote.controller.StaticFileController;
+import org.apache.coyote.http11.request.Request;
 import org.apache.coyote.http11.URL;
 import org.junit.jupiter.api.Test;
