+++ tomcat/src/test/java/nextstep/org/apache/coyote/http11/Http11ProcessorTest.java
@@ -11,8 +11,8 @@ import java.net.URL;
 import java.nio.file.Files;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import org.apache.coyote.HttpMethod;
-import org.apache.coyote.HttpStatus;
+import org.apache.coyote.http11.request.HttpMethod;
+import org.apache.coyote.http11.HttpStatus;
 import org.apache.coyote.http11.Http11Processor;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
