+++ tomcat/src/test/java/nextstep/org/apache/coyote/http11/QueryParamsTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import nextstep.jwp.exception.InvalidQueryParamKeyException;
-import org.apache.coyote.http11.QueryParams;
+import org.apache.coyote.http11.request.QueryParams;
 import org.junit.jupiter.api.Test;
 
 class QueryParamsTest {
