+++ tomcat/src/test/java/nextstep/org/apache/coyote/http11/ResponseTest.java
@@ -6,8 +6,8 @@ import java.io.File;
 import java.io.IOException;
 import java.net.URISyntaxException;
 import java.nio.file.Files;
-import org.apache.coyote.HttpMethod;
-import org.apache.coyote.HttpStatus;
+import org.apache.coyote.http11.request.HttpMethod;
+import org.apache.coyote.http11.HttpStatus;
 import org.apache.coyote.http11.Response;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
